add dma region to the 54h20 radio overlay
While this will be fixed in the near future by amending the default
device settings, we would like to have this working sooner.
Without this the radio core will just "die" without any message and will
need to be recovered - and the sample (when build for the radio core)
doesn't work with a dead netcore.

Signed-off-by: Kyra Lengfeld <kyra.lengfeld@nordicsemi.no>